requirements.txt
argcomplete==1.12.2
boto3==1.17.33
botocore==1.20.33
cachetools==4.2.1
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
durationpy==0.5
future==0.18.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
hjson==3.0.2
httplib2==0.19.0
idna==2.10
importlib-metadata==3.7.3
jmespath==0.10.0
kappa==0.6.0
oauthlib==3.1.0
packaging==20.9
pep517==0.10.0
pip-tools==6.0.1
placebo==0.9.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.6
six==1.15.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.59.0
troposphere==2.6.4
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.4.1
